client message UUid can now be string or byte[]
The UUid generation for the client is now a function called by the client.  By default, the function used is ZeroRpc.Net.Util.UuidGen.ComputeUuid. But you can pass ZeroRpc.Net.Util.UuidGen.ComputeUuidByteArray or even your own function.
An new IEqualityComparer has been implemented in ChannelKeyComparer in order to be able to find the channel in dictionnary Dictionary<object, Channel>, in the case of byte[] UUId.